Disable `translate3D` for list when scaling (fixes #24599)
