+++ src/vs/base/browser/browser.ts
@@ -124,7 +124,34 @@ export const isChrome = (userAgent.indexOf('Chrome') >= 0);
 export const isSafari = (userAgent.indexOf('Chrome') === -1) && (userAgent.indexOf('Safari') >= 0);
 export const isIPad = (userAgent.indexOf('iPad') >= 0);
 
-export const canUseTranslate3d = !isFirefox;
+const isChromev56 = (
+	navigator.userAgent.indexOf('Chrome/56.') >= 0
+	// Edge likes to impersonate Chrome sometimes
+	&& navigator.userAgent.indexOf('Edge/') === -1
+);
+
+export const supportsTranslate3d = !isFirefox;
+
+export function canUseTranslate3d(): boolean {
+	if (!supportsTranslate3d) {
+		return false;
+	}
+
+	if (getZoomLevel() !== 0) {
+		return false;
+	}
+
+	// see https://github.com/Microsoft/vscode/issues/24483
+	if (isChromev56) {
+		const pixelRatio = getPixelRatio();
+		if (Math.floor(pixelRatio) !== pixelRatio) {
+			// Not an integer
+			return false;
+		}
+	}
+
+	return true;
+}
 
 export const enableEmptySelectionClipboard = isWebKit;
